feat(Observables): Change the API to implement Symbol.observable forcing use of Observable.from to b
This takes away the subscribe method, and forces the use of an Observable.from API like RxJS and the
ECMA observable spec has. This leaves the creation of the event streams up to the user and whatever
event stream library they use.

BREAKING CHANGE: Removed the subscribe method in faavour of Symbol.observable on instances